I have updated the Code into MVC. everything is working normal but there is a issue in the challenges module at completeChallnege function. it happens because of the calculaetpoints cant properly imported
