+++ package.json
@@ -14,6 +14,8 @@
     "express": "^4.21.1",
     "express-session": "^1.18.1",
     "mongoose": "^8.8.2",
-    "multer": "^1.4.5-lts.1"
+    "multer": "^1.4.5-lts.1",
+    "nodemon": "^3.1.9",
+    "zod": "^3.24.2"
   }
 }
